test for evaluator #3

+++ commons/src/main/java/de/hpi/mmds/wiki/Evaluator.java
@@ -2,7 +2,6 @@ package de.hpi.mmds.wiki;
 
 import org.apache.commons.io.FileUtils;
 import org.apache.spark.api.java.JavaRDD;
-import org.apache.spark.api.java.JavaSparkContext;
 
 import java.io.BufferedWriter;
 import java.io.File;
@@ -19,6 +18,38 @@ public class Evaluator {
 	private final File out;
 	private final Recommender recommender;
 
+	public static class Result<T> {
+
+		private final Set<T> recommendations;
+		private final Set<T> groundTruth;
+		private final Set<T> intersect;
+
+		public Result(Set<T> recommendations, Set<T> groundTruth) {
+			this.recommendations = recommendations;
+			this.groundTruth = groundTruth;
+			this.intersect = new HashSet<>(groundTruth);
+			this.intersect.retainAll(recommendations);
+		}
+
+		public double precision() {
+			return recommendations.isEmpty() ? 0 : (double) intersect.size() / recommendations.size();
+		}
+
+		public double recall() {
+			return (double) intersect.size() / groundTruth.size();
+		}
+
+		public String printResult() {
+			StringBuilder sb = new StringBuilder();
+			sb.append("Recommendations: " + recommendations + "\n");
+			sb.append("Gold standard: " + groundTruth + "\n");
+			sb.append("Matches: " + intersect + "\n");
+			sb.append("Precision: " + precision() + "\n");
+			sb.append("Recall: " + recall());
+			return sb.toString();
+		}
+	}
+
 	public Evaluator(Recommender recommender, Edits test, Edits training, File out) {
 		this.recommender = recommender;
 		this.training = training;
@@ -26,18 +57,21 @@ public class Evaluator {
 		this.out = out;
 	}
 
-	public void evaluate(int num, long seed) {
+	public Map<Integer, Result> evaluate(int num, long seed) {
+		Map<Integer, Result> results = new HashMap<>();
 		List<Integer> uids = new ArrayList<>(test.getUsers().intersection(training.getUsers()).collect());
 		int i = 0;
 		double totalPrecision = 0.0;
 		double totalRecall = 0.0;
-		try {
-			FileUtils.forceDelete(out);
-		} catch (IOException e) {
-			throw new RuntimeException("Could not delete output file " + out.getPath(), e);
+		if (out.exists()) {
+			try {
+				FileUtils.forceDelete(out);
+			} catch (IOException e) {
+				throw new RuntimeException("Could not delete output file " + out.getPath(), e);
+			}
 		}
 		Collections.shuffle(uids, new Random(seed));
-		try (BufferedWriter writer = new BufferedWriter(new FileWriter(this.out))) {
+		try (BufferedWriter writer = new BufferedWriter(new FileWriter(out))) {
 			for (int user : uids) {
 				if (i >= num) {
 					break;
@@ -48,27 +82,21 @@ public class Evaluator {
 				Set<Integer> groundTruth = new HashSet<>(a);
 				groundTruth.removeAll(p);
 				if (!groundTruth.isEmpty()) {
-					List<Integer> recommendations = recommender.recommend(user, articles, NUM_RECOMMENDATIONS).stream()
-							.map(Recommendation::getArticle).collect(Collectors.toList());
-					Set<Integer> intersect = new HashSet<>(groundTruth);
-					intersect.retainAll(recommendations);
-					double precision = recommendations.isEmpty() ?
-							0 :
-							(double) intersect.size() / recommendations.size();
-					double recall = (double) intersect.size() / groundTruth.size();
-					totalPrecision += precision;
-					totalRecall += recall;
+					Set<Integer> recommendations = recommender.recommend(user, articles, NUM_RECOMMENDATIONS).stream()
+							.map(Recommendation::getArticle).collect(Collectors.toSet());
+					Result<Integer> result = new Result<>(recommendations, groundTruth);
+					results.put(user, result);
+					totalPrecision += result.precision();
+					totalRecall += result.recall();
 					i++;
 					writer.write("User: " + user + "\n");
-					writer.write("Recommendations: " + recommendations + "\n");
-					writer.write("Gold standard: " + groundTruth + "\n");
-					writer.write("Matches: " + intersect + "\n");
-					writer.write("Precision: " + precision + "\n");
+					writer.write(result.printResult());
+					writer.newLine();
 					writer.write("AVG Precision: " + totalPrecision / i + "\n");
-					writer.write("Recall: " + recall + "\n");
 					writer.write("AVG Recall: " + totalRecall / i + "\n");
 					writer.write("Processed: " + i + "\n");
-					writer.write("---\n");
+					writer.write("---");
+					writer.newLine();
 					writer.flush();
 				}
 			}
@@ -79,9 +107,10 @@ public class Evaluator {
 			System.out.println("AVG Precision: " + totalPrecision / i);
 			System.out.println("AVG Recall: " + totalRecall / i);
 		}
+		return results;
 	}
 
-	public void evaluate(int num) {
-		evaluate(num, 1L);
+	public Map<Integer, Result> evaluate(int num) {
+		return evaluate(num, 1L);
 	}
 }
